requirements.txt
Flask~=3.0.3
python-jobspy~=1.1.65
numpy~=2.1.1
pandas~=2.2.2
python_dateutil~=2.9.0.post0
Requests~=2.32.3
scikit_learn~=1.5.2
sentence_transformers~=3.1.1
torch~=2.4.1
transformers~=4.45.1

redis~=5.1.0
aiohttp~=3.10.5
python-magic~=0.4.27
pdfplumber~=0.11.4
python-docx~=1.1.2
pypandoc~=1.13
Werkzeug~=3.0.4
pytest~=8.3.3
selenium~=4.25.0
webdriver-manager~=4.0.2
reportlab~=4.2.2